Read uint16_t for fixedRings

+++ include/file_tools.h
@@ -2,10 +2,11 @@
 #define FILE_TOOLS_H
 
 #include "types.h"
+#include <cstdint>
 #include <filesystem>
 #include <unordered_set>
 
-std::unordered_set<int> readFixedRings(const std::filesystem::path &filePath,
-                                       const LoggerPtr &logger);
+std::unordered_set<uint16_t>
+readFixedRings(const std::filesystem::path &filePath, const LoggerPtr &logger);
 
 #endif // FILE_TOOLS_H
